Consolidate iOS stored payment deletion ownership
DropInPlatformApi already coordinates the deletion request and result, so conform directly to StoredPaymentMethodsDelegate and remove the pass-through delegate.
